Split out the analysis package.
