+++ src/main/scala/com/cibo/scalastan/analysis/LoopChecker.scala
@@ -8,10 +8,11 @@
  * or at https://opensource.org/licenses/BSD-3-Clause
  */
 
-package com.cibo.scalastan.transform
+package com.cibo.scalastan.analysis
 
 import com.cibo.scalastan.ScalaStan
 import com.cibo.scalastan.ast._
+import com.cibo.scalastan.transform.StanTransform
 
 // Make sure all "break" and "continue" statements are in loops.
 class LoopChecker(implicit val ss: ScalaStan) extends StanTransform {
