+++ src/main/scala/com/cibo/scalastan/analysis/UseDefinitions.scala
@@ -1,4 +1,4 @@
-package com.cibo.scalastan.transform
+package com.cibo.scalastan.analysis
 
 import com.cibo.scalastan.ast.{StanProgram, StanStatement}
